+++ packages/datadog_tracking_http_client/lib/src/tracking_http_client_plugin.dart
@@ -65,8 +65,12 @@ class _DdHttpTrackingPlugin extends DatadogPlugin {
 
   @override
   void initialize() {
-    HttpOverrides.global =
-        DatadogTrackingHttpOverrides(instance, configuration);
+    final existingOverrides = HttpOverrides.current;
+    HttpOverrides.global = DatadogTrackingHttpOverrides(
+      instance,
+      configuration,
+      existingOverrides,
+    );
     instance.updateConfigurationInfo(
         LateConfigurationProperty.trackNetworkRequests, true);
   }